Fixed dtype error in Atlantis

+++ atmod/preprocessing/process_soilmap.py
@@ -8,7 +8,7 @@ from atmod.warnings import suppress_warnings
 
 EMPTYDICT = numba.typed.Dict.empty(
     key_type=numba.types.int16,
-    value_type=numba.types.float32[:],
+    value_type=numba.types.float64[:],
 )
 
 
@@ -64,10 +64,10 @@ class NumbaDicts:
             org = cls._get_values(df['orgmatter'] / to_fraction, ascending_depth)
             lut = cls._get_values(df['lutum'] / to_fraction, ascending_depth)
 
-            lithology[nr] = np.float32(lith)
-            thickness[nr] = np.float32(thick)
-            organic[nr] = np.float32(org)
-            lutum[nr] = np.float32(lut)
+            lithology[nr] = np.float64(lith)
+            thickness[nr] = np.float64(thick)
+            organic[nr] = np.float64(org)
+            lutum[nr] = np.float64(lut)
 
         return cls(thickness, lithology, organic, lutum)
